+++ frontend/src/js/charts/project.js
@@ -1,0 +1,71 @@
+//@ts-check
+/// <reference path="../types/index.d.ts" />
+
+let {
+	convertUnit2Hour,
+	orderByWatchingTime,
+	object2array,
+	getEachFieldToFixed2,
+	getShortProjectName,
+} = require('../utils'), {
+	createEChartsSeries,
+	AXIS_HOURS,
+	GRID_NORMAL
+} = require('../echartsUtils');
+
+const SELECTOR = '#chartProject';
+const SIZE = 5;
+
+/**
+* @type {string[]}
+*/
+let originalProjectNames = [], projectNames = [], shortProjectNames = [];
+
+function tooltipFormatter(p, ticket, set) {
+	let setText = text => (setTimeout(set, 1, ticket, text), text);
+	let i = p.dataIndex;
+	if (i >= projectNames.length) return setText(null);
+	return setText(`You spent<br/> (<b>${p.value}</b> hours)<br/> on <u>${projectNames[i]}</u>`);
+}
+
+/**
+ * @type {EChartsObject}
+ */
+let charts = null;
+
+module.exports = { update };
+
+function update(dataGroupByProject) {
+	if (!charts) {
+		charts = echarts.init($(SELECTOR)[0]);
+		charts.on('click', params =>
+			typeof params.dataIndex == 'number' &&
+			global.app.openProjectReport(originalProjectNames[params.dataIndex]));
+	}
+
+	let data = convertUnit2Hour(dataGroupByProject),
+		array = orderByWatchingTime(object2array(data)).slice(-SIZE);
+	originalProjectNames = array.map(it => it.name);
+	projectNames = array.map(it => decodeURIComponent(it.name));
+	shortProjectNames = projectNames.map((name, i) =>
+		getShortProjectName(name) + ` (${Number(array[i].watching).toFixed(2)} hs)`);
+
+	charts.setOption({
+		legend: { data: [''] },
+		xAxis: {
+			type: 'value', nameLocation: 'end', position: 'top',
+			axisTick: { show: false }, axisLabel: AXIS_HOURS.axisLabel },
+		yAxis: {
+			type: 'category', nameLocation: 'start',
+			axisTick: { show: false }, axisLabel: { inside: true }, z: 1024,
+			data: shortProjectNames },
+		grid: GRID_NORMAL,
+		tooltip: { trigger: 'item', formatter: tooltipFormatter},
+		series: [
+			createEChartsSeries('bar', 'watching')
+				.setItemColor('#fff59d')
+				.setValues(getEachFieldToFixed2(array ,'watching'))
+				.toObject()
+		]
+	});
+}